+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
   this allows to specify how taint propagates through side-effectful function calls.
   For example, you can specify that when tainted data is added to an array then the
   array itself becomes tainted. (#4509)
+- Add `--github-actions` option for GitHub Actions workflow commands format.
 
 ### Changed